+++ Shared/Helpers.swift
@@ -27,7 +27,6 @@ extension Color {
     
 }
 
-
 protocol PositionInSuperView {
     
     var positionInSuperView: CGPoint {
