+++ iOS/SwiftUIMetalKitView.swift
@@ -13,29 +13,3 @@ class Setting {
         self.bool = newBool
     }
 }
-
-
-struct SwiftUIMetalKitView: UIViewControllerRepresentable {
-    
-    typealias UIViewControllerType = MetalBrotViewController
-    typealias UIViewType = UIView
-    
-    func makeUIViewController(context: Context) -> MetalBrotViewController {
-        MetalBrotViewController()
-    }
-    
-    func updateUIViewController(_ uiViewController: MetalBrotViewController, context: Context) {
-        print("view updated")
-    }
-    
-    func makeCoordinator() -> Setting {
-        Setting()
-    }
-    
-}
-
-struct SwiftUIMetalKitView_Previews: PreviewProvider {
-    static var previews: some View {
-        SwiftUIMetalKitView()
-    }
-}
