+++ macOS/MetalbrotViewController-macOS.swift
@@ -18,7 +18,7 @@ class MyView: NSView {
     }
 }
 
-class ViewController: NSViewController {
+class MetalbrotViewController: NSViewController {
 
     var metalView: MTKView!
     var renderer: MetalbrotRenderer?
@@ -60,11 +60,6 @@ class ViewController: NSViewController {
 
     }
     
-//    override func viewDidLayout() {
-//        print("view center: \(self.view.bounds.center) \(self.view.frame.center)")
-//        translation = self.view.frame.center
-//    }
-    
     override func mouseDragged(with event: NSEvent) {
         
         let translation = CGPoint(x: translation.x - event.deltaX, y: translation.y - event.deltaY)
@@ -88,31 +83,6 @@ class ViewController: NSViewController {
     
 }
 
-extension ViewController: MetalViewUpdateDelegate {
-    
-    func translationDidUpdate(point: CGPoint) {
-        print("got updated translation \(point)")
-        translation = point
-    }
-    
-}
 
-struct SwiftUIMetalKitView: NSViewControllerRepresentable {
 
-    typealias NSViewControllerType = ViewController
-    typealias NSViewType = NSView
-    
-    func makeNSViewController(context: Context) -> ViewController {
-        ViewController()
-    }
-    
-    func updateNSViewController(_ nsViewController: ViewController, context: Context) {
-        print("view updated")
-    }
-    
-    func makeCoordinator() -> Bool? {
-        true
-    }
-    
-}
